Use placeholders for credentails

+++ src/main/java/com/testing/geologicalsectionsapi/services/impl/SectionServiceImpl.java
@@ -1,12 +1,14 @@
 package com.testing.geologicalsectionsapi.services.impl;
 
 import com.testing.geologicalsectionsapi.exception.CustomRequestException;
+import com.testing.geologicalsectionsapi.models.GeologicalClass;
 import com.testing.geologicalsectionsapi.models.Section;
 import com.testing.geologicalsectionsapi.repositories.SectionRepository;
 import com.testing.geologicalsectionsapi.services.SectionService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class SectionServiceImpl implements SectionService {
@@ -18,14 +20,11 @@ public class SectionServiceImpl implements SectionService {
         this.sectionRepository = sectionRepository;
     }
 
-
-
-    public Section getSectionById(Long id, String traceId) {
-        return sectionRepository.findById(id)
-                .orElseThrow(() -> new CustomRequestException(traceId));
+    @Override
+    public Optional<Section> getSectionById(Long id, String traceId) {
+        return sectionRepository.findById(id);
 
     }
-
 
     @Override
     public Section createSection(Section section) {
@@ -43,9 +42,16 @@ public class SectionServiceImpl implements SectionService {
         return sectionRepository.save(existingSection);
     }
 
+
     @Override
-    public void deleteSection(Long id) {
-        sectionRepository.deleteById(id);
+    public boolean deleteSection(Long id) {
+        // Check if the class exists before deleting
+        if (sectionRepository.existsById(id)) {
+            sectionRepository.deleteById(id);
+            return true; // Deletion successful
+        } else {
+            return false; // Class with the given ID not found
+        }
     }
